Added Proxy backend class and auto-detection of AWS SDK versions
