+++ test/PHPQueue/Backend/AmazonSQSV1Test.php
@@ -1,6 +1,7 @@
 <?php
 namespace PHPQueue\Backend;
-class AmazonSQSTest extends \PHPUnit_Framework_TestCase
+
+class AmazonSQSV1Test extends \PHPUnit_Framework_TestCase
 {
     private $object;
 
@@ -11,15 +12,16 @@ class AmazonSQSTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('Amazon PHP SDK not installed');
         } else {
             $options = array(
-                  'region'      => \AmazonSQS::REGION_APAC_SE1
-                , 'queue'       => 'https://sqs.ap-southeast-1.amazonaws.com/524787626913/testqueue'
-//                , 'sqs_options' => array(
-//                                      'key'    => 'xxx'
-//                                    , 'secret' => 'xxx'
-//                                )
-                , 'receiving_options' => array('VisibilityTimeout' => 0)
+                'region' => \AmazonSQS::REGION_APAC_SE1,
+                'queue' => 'https://sqs.ap-southeast-1.amazonaws.com/524787626913/testqueue',
+                'sqs_options' => array(
+                    'key' => 'your_sqs_key',
+                    'secret' => 'your_sqs_secret'
+                ),
+                'receiving_options' => array('VisibilityTimeout' => 0)
             );
-            $this->object = new AmazonSQS($options);
+            $this->object = new AmazonSQS();
+            $this->object->setBackend(new Aws\AmazonSQSV1($options));
         }
     }
